Improve priorities
